requirements.txt
black==21.5b1
Flask==1.1.2
grpcio==1.32.0
gunicorn==20.1.0
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.4
Pillow==8.2.0
pytest==6.2.4
scipy==1.6.3
seaborn==0.11.1
tensorflow==2.4.1
tensorflow-io==0.17.1
tensorflow-serving-api==2.4.1